flutter util refactor
