+++ lib/src/util/flutter_utils.dart
@@ -3,6 +3,7 @@
 // BSD-style license that can be found in the LICENSE file.
 
 import 'package:analyzer/dart/element/element.dart';
+import 'package:analyzer/dart/element/nullability_suffix.dart';
 import 'package:analyzer/dart/element/type.dart';
 
 import '../util/dart_type_utilities.dart';
@@ -19,6 +20,12 @@ _Flutter _flutterInstance = _Flutter('flutter', 'package:flutter');
 
 _Flutter get _flutter => _flutterInstance;
 
+bool hasWidgetAsAscendant(ClassElement element) =>
+    _flutter.hasWidgetAsAscendant(element);
+
+bool isBuildContext(DartType? type, {bool skipNullable = false}) =>
+    _flutter.isBuildContext(type, skipNullable: skipNullable);
+
 bool isExactWidget(ClassElement element) => _flutter.isExactWidget(element);
 
 bool isExactWidgetTypeContainer(DartType? type) =>
@@ -26,9 +33,6 @@ bool isExactWidgetTypeContainer(DartType? type) =>
 
 bool isStatefulWidget(ClassElement? element) =>
     _flutter.isStatefulWidget(element);
-
-bool hasWidgetAsAscendant(ClassElement element) =>
-    _flutter.hasWidgetAsAscendant(element);
 
 bool isWidgetProperty(DartType? type) {
   if (isWidgetType(type)) {
@@ -46,6 +50,7 @@ bool isWidgetType(DartType? type) => _flutter.isWidgetType(type);
 
 /// See: analysis_server/lib/src/utilities/flutter.dart
 class _Flutter {
+  static const _nameBuildContext = 'BuildContext';
   static const _nameStatefulWidget = 'StatefulWidget';
   static const _nameWidget = 'Widget';
   static const _nameContainer = 'Container';
@@ -61,12 +66,37 @@ class _Flutter {
         _uriContainer = Uri.parse('$uriPrefix/src/widgets/container.dart'),
         _uriFramework = Uri.parse('$uriPrefix/src/widgets/framework.dart');
 
-  bool isExactWidgetTypeContainer(DartType? type) =>
-      type is InterfaceType &&
-      _isExactWidget(type.element, _nameContainer, _uriContainer);
+  bool hasWidgetAsAscendant(ClassElement? element,
+      [Set<ClassElement>? alreadySeen]) {
+    alreadySeen ??= {};
+    if (element == null || alreadySeen.contains(element)) {
+      return false;
+    }
+    alreadySeen.add(element);
+    if (_isExactWidget(element, _nameWidget, _uriFramework)) {
+      return true;
+    }
+    return hasWidgetAsAscendant(element.supertype?.element, alreadySeen);
+  }
+
+  bool isBuildContext(DartType? type, {bool skipNullable = false}) {
+    if (type is! InterfaceType) {
+      return false;
+    }
+    if (skipNullable && type.nullabilitySuffix == NullabilitySuffix.question) {
+      return false;
+    }
+    var element = type.element;
+    return element.name == _nameBuildContext &&
+        element.source.uri == _uriFramework;
+  }
 
   bool isExactWidget(ClassElement element) =>
       _isExactWidget(element, _nameWidget, _uriFramework);
+
+  bool isExactWidgetTypeContainer(DartType? type) =>
+      type is InterfaceType &&
+      _isExactWidget(type.element, _nameContainer, _uriContainer);
 
   bool isStatefulWidget(ClassElement? element) {
     if (element == null) {
@@ -93,19 +123,6 @@ class _Flutter {
       }
     }
     return false;
-  }
-
-  bool hasWidgetAsAscendant(ClassElement? element,
-      [Set<ClassElement>? alreadySeen]) {
-    alreadySeen ??= {};
-    if (element == null || alreadySeen.contains(element)) {
-      return false;
-    }
-    alreadySeen.add(element);
-    if (_isExactWidget(element, _nameWidget, _uriFramework)) {
-      return true;
-    }
-    return hasWidgetAsAscendant(element.supertype?.element, alreadySeen);
   }
 
   bool isWidgetType(DartType? type) =>
